django/requirements.txt
Django
djangorestframework
djangorestframework-jwt
psycopg2